pyproject.toml
[project]
name = "agent-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "strands-agents[ollama]>=1.36.0",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.5.0",
    "gTTS>=2.3.0",
    "pydantic_settings>=2.0.0",
]
